Move the agents onto a configurable LLM provider
The Cerebras key had no billing, so every agent call returned 402 or 404 and
the AI features had never once succeeded. Point the agents at any
OpenAI-compatible endpoint via LLM_BASE_URL/LLM_MODEL instead of hardcoding a
provider and model in three files, so the next rotation is a config change.

Working through the failures uncovered four bugs the dead LLM had been hiding:

- Postgres NUMERIC arrives as Decimal, and building numpy arrays without an
  explicit dtype produced object arrays, so RSI died mixing Decimal with float.
  The MCP tool swallowed that, returned no indicators, and the indicator agent
  had been analysing nothing at all.
- Both agents returned the model's echoed JSON wholesale, letting invented or
  null indicator values and price levels replace real computed ones. Strategy
  logic could act on hallucinated RSI. Computed values now win; the model
  contributes synthesis and narrative only.
- TradingStrategy got a duplicate symbol kwarg and had no default created_at.
- LiquidationLevel.type is required but was never populated; derive it from
  which list a level is in rather than trusting the model for it.

+++ backend/services/market_data_service.py
@@ -87,10 +87,12 @@ class MarketDataService:
         if not ohlc_data:
             return {}
 
-        closes = np.array([candle['close'] for candle in ohlc_data])
-        highs = np.array([candle['high'] for candle in ohlc_data])
-        lows = np.array([candle['low'] for candle in ohlc_data])
-        volumes = np.array([candle['volume'] for candle in ohlc_data])
+        # Postgres NUMERIC arrives as Decimal; without an explicit dtype numpy
+        # builds object arrays and the indicator math dies mixing Decimal with float.
+        closes = np.array([candle['close'] for candle in ohlc_data], dtype=float)
+        highs = np.array([candle['high'] for candle in ohlc_data], dtype=float)
+        lows = np.array([candle['low'] for candle in ohlc_data], dtype=float)
+        volumes = np.array([candle['volume'] for candle in ohlc_data], dtype=float)
 
         indicators = {}
